// frontend/src/api/models/SchemaUserVideoCard.ts
export interface SchemaUserVideoCard {
    /**
     * 
     * @type {number}
     * @memberof SchemaUserVideoCard
     */
    videoDuration: number;
    /**
     * 
     * @type {VideoAspectRatio}
     * @memberof SchemaUserVideoCard
     */
    videoAspectRatio: VideoAspectRatio;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserVideoCard
     */
    videoTitle: string;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserVideoCard
     */
    videoId: string;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserVideoCard
     */
    videoThumbnailUrl: string;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserVideoCard
     */
    videoUrl: string;
    /**
     * 
     * @type {CreatedAt}
     * @memberof SchemaUserVideoCard
     */
    createdAt: CreatedAt;
    /**
     * 
     * @type {UpdatedAt}
     * @memberof SchemaUserVideoCard
     */
    updatedAt: UpdatedAt;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserVideoCard
     */
    status: string;
    /**
     * 
     * @type {AppDbModelDocumentVideoProcessingApplication}
     * @memberof SchemaUserVideoCard
     */
    appliedApplication?: AppDbModelDocumentVideoProcessingApplication;
}